Add files via upload
Added the first version of a python script for making change.  The bulk of the code was taken from a wikipedia article, but I added in a second matrix to keep track of the optimal denominations actually used to make the amount.  Thus, a table can be generated showing the optimum set of denominations at each value.  I hope to eventually use this for doing some research in optimization.